% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{buoy_tsibble}
\alias{buoy_tsibble}
\title{Convert a buoy tibble to a tsibble}
\usage{
buoy_tsibble(x)
}
\arguments{
\item{x}{tibble of buoy data}
}
\value{
tsibble
}
\description{
Convert a buoy tibble to a tsibble
}
