% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{diff_time}
\alias{diff_time}
\title{Compute difftime on a vector in user specified units}
\usage{
diff_time(x, units = c("secs", "mins", "hours", "days")[1])
}
\arguments{
\item{x}{POSIXct time vector}

\item{units}{character, one of "secs" (default), "mins", "hours", "days"}
}
\value{
\code{difftime} object with specified units
}
\description{
Compute difftime on a vector in user specified units
}
\seealso{
\href{https://stackoverflow.com/questions/40470389/diff-for-posixct-with-units-like-in-difftime}{stackoverflow post}
}
