% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds.R
\name{get_catalog}
\alias{get_catalog}
\title{Given a catalog uri return a \code{\link[thredds]{CatalogNode}}}
\usage{
get_catalog(
  program = "University of Maine Buoys",
  uri = thredds_base_url(what = "xml")
)
}
\arguments{
\item{program}{character, the name of the program, if not found
then the top level catalog is returned}

\item{uri}{character, the url for the catalog node}
}
\value{
\code{\link[thredds]{CatalogNode}}
}
\description{
Given a catalog uri return a \code{\link[thredds]{CatalogNode}}
}
