% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds.R
\name{get_Maine_buoy}
\alias{get_Maine_buoy}
\title{Get Maine a Maine buoy reference}
\usage{
get_Maine_buoy(
  suite = "Historic and Realtime Aggregations",
  buoy = "A01",
  dataset = "CTD 1m",
  form = c("url", "ncdf")[1]
)
}
\arguments{
\item{suite}{character, one of these \url{http://www.neracoos.org/thredds/UMO_all.html}{UMO_all}}

\item{buoy}{character, one of the buoys listed here \url{http://www.neracoos.org/thredds/UMO_historical_realtime_agg.html}{Historic/Realtime Aggregations}}

\item{dataset}{character, one of the instruments listed here \url{http://www.neracoos.org/thredds/UMO_historical_realtime_agg.html}{datasets}}

\item{form}{character, either "url" or "ncdf" to specify the type of output}
}
\value{
either a character URL to a dataset, ncdf4 class dataset object, or NULL
}
\description{
Get Maine a Maine buoy reference
}
