% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_close}
\alias{nc_close}
\title{Close an `ncdf4` connection}
\usage{
nc_close(x)
}
\arguments{
\item{x}{ncdf4 class object to close}
}
\description{
Close an `ncdf4` connection
}
\seealso{
\code{\link[ncdf4]{nc_close}}
}
