% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_get_dim}
\alias{nc_get_dim}
\title{Retrieve a dimension by name}
\usage{
nc_get_dim(x, name = "lon")
}
\arguments{
\item{x}{ncdf4 class object}

\item{name}{the name of the dimension to retrieve}
}
\value{
vector of dimension values
}
\description{
Retrieve a dimension by name
}
