% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_get_time}
\alias{nc_get_time}
\title{Retrieve the time dimension from a ncdf4 class object}
\usage{
nc_get_time(x)
}
\arguments{
\item{x}{ncdf4 class object}
}
\value{
numerc vector of POSIXct
}
\description{
Retrieve the time dimension from a ncdf4 class object
}
