% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_get_var}
\alias{nc_get_var}
\title{Retrieve a variable by name}
\usage{
nc_get_var(x, name = "salinity", ...)
}
\arguments{
\item{x}{ncdf4 class object}

\item{name}{the name of the variable to retrieve}

\item{...}{other arguments for \code{\link[ncdf4]{ncvar_get}}}
}
\value{
vector of variable values
}
\description{
Retrieve a variable by name
}
