% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_has_dim}
\alias{nc_has_dim}
\title{Test if ncdf4 object has a dimension}
\usage{
nc_has_dim(x, name = "time")
}
\arguments{
\item{x}{ncdf4 class object}

\item{name}{character, name of the dimension}
}
\value{
logical TRUE if the dimension is present
}
\description{
Test if ncdf4 object has a dimension
}
