% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_has_var}
\alias{nc_has_var}
\title{Test if ncdf4 object has a variable}
\usage{
nc_has_var(x, name = "foo")
}
\arguments{
\item{x}{ncdf4 class object}

\item{name}{character, name of the variable}
}
\value{
logical TRUE if the variable is present
}
\description{
Test if ncdf4 object has a variable
}
