% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{regularize}
\alias{regularize}
\title{Regularize an irregular tsibble using the specified interval and function}
\usage{
regularize(x, unit = "30 min", fun = lubridate::round_date)
}
\arguments{
\item{x}{irregular tsibble}

\item{unit}{character, the regularization interval, by default "30 min"}

\item{fun}{function, the name of the function to apply for the regularization, by default `lubridate::date_round`}
}
\value{
regularized tsibble
}
\description{
Regularize an irregular tsibble using the specified interval and function
}
\seealso{
\code{\link[tsibble]{index_by}}

\code{\link[lubridate]{round_date}}

\href{https://blog.earo.me/2018/12/20/reintro-tsibble/#new-time-based-verbs}{blog-post}
}
