% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{nc_get_table}
\alias{nc_get_table}
\title{Retrieve a table of dimensions and variables by name.  Dimensions
and variables must have same length.}
\usage{
nc_get_table(
  x,
  dnames = "time",
  vnames = c("conductivity", "temperature", "salinity", "sigma_t"),
  ...
)
}
\arguments{
\item{x}{ncdf4 class object}

\item{dnames}{character, the name of the dimensions to retrieve}

\item{vnames}{character, the name of the variables to retrieve}

\item{...}{other arguments for \code{nc_get_var}}
}
\value{
table of dimensions and variables
}
\description{
Retrieve a table of dimensions and variables by name.  Dimensions
and variables must have same length.
}
